invoice pdf generation in progress

+++ app/Domain/Model/Documents/Quote/QuoteRepository.php
@@ -30,6 +30,7 @@ class QuoteRepository extends AbstractDocumentRepository
             'po_number' => $data['po_number'],
             'partial' => $data['partial'],
             'discount' => $data['discount_value'],
+            'currency_id' => $data['currency_id'],
             'discount_type' => $data['discount_type'],
             'date' => $data['quote_date'],
             'due_date' => $data['due_date'],
@@ -52,30 +53,40 @@ class QuoteRepository extends AbstractDocumentRepository
 
     public function updated(&$quote, &$data, &$protectedData)
     {
-        $quote->bill->update([
-            'number' => $data['quote_number'],
-            'po_number' => $data['po_number'],
-            'partial' => $data['partial'],
-            'discount' => $data['discount_value'],
-            'discount_type' => $data['discount_type'],
-            'date' => $data['quote_date'],
-            'due_date' => $data['due_date'],
-            'notes' => $data['note_to_client'],
-            'terms' => $data['terms'],
-            'footer' => $data['footer']
-        ]);
+        $billData = [];
 
-        $quote->bill->items()->delete();
+        foreach ([
+            'number' => 'number',
+            'po_number' => 'po_number',
+            'partial' => 'partial',
+            'discount' => 'discount',
+            'discount_type' => 'discount_type',
+            'currency_id' => 'currency_id',
+            'date' => 'date',
+            'due_date' => 'due_date',
+            'notes' => 'notes',
+            'terms' => 'terms',
+            'footer' => 'footer'
+        ] as $field => $billField) {
+            if (array_key_exists($field, $data)) {
+                $billData[$billField] = $data[$field];
+            }
+        }
+        $quote->bill->update($billData);
 
-        foreach ($data['items'] as $index => $item) {
-            $quote->bill->items()->create([
-                'product_uuid' => $item['product_uuid'],
-                'cost' => $item['cost'],
-                'qty' => $item['qty'],
-                'discount' => $item['discount'],
-                'tax_rate_uuid' => $item['tax_rate_uuid'],
-                'index' => $index
-            ]);
+        if (isset($data['items'])) {
+            $quote->bill->items()->delete();
+
+            foreach ($data['items'] as $index => $item) {
+                $quote->bill->items()->create([
+                    'product_uuid' => $item['product_uuid'],
+                    'cost' => $item['cost'],
+                    'qty' => $item['qty'],
+                    'discount' => $item['discount'],
+                    'tax_rate_uuid' => $item['tax_rate_uuid'],
+                    'index' => $index
+                ]);
+            }
         }
 
         return $quote;
